+++ utils.h
@@ -14,5 +14,5 @@ using namespace std;
 
 void printMatrixValues(vector<vector<int>> matrix);
 vector<vector<int>> readMatrixFromTxtFile(string matrix_filename);
-void saveMatrixOnTxtFile(vector<vector<int>> matrix, int dim_n, int dim_m, int time_result, string file_name, string algorithm);
+void saveMatrixOnTxtFile(vector<vector<int>> matrix, int dim_n, int dim_m, double time_result, string file_name, string algorithm);
 void multiplyMatricesSequential(vector<vector<int>> matrix_one, vector<vector<int>> matrix_two);
